+++ cli/cli-core/src/main/java/io/telicent/smart/cache/cli/commands/projection/AbstractProjectorCommand.java
@@ -52,7 +52,7 @@ import java.util.function.Supplier;
  */
 // java:S119 - TKey/TValue/TRequest generic naming convention is used across the codebase
 // java:S4507 - stack trace is intentionally written to stderr for CLI diagnostics
-@SuppressWarnings({"java:S119", "java:S4507"})
+@SuppressWarnings({ "java:S119", "java:S4507" })
 public abstract class AbstractProjectorCommand<TKey, TValue, TOutput> extends SmartCacheCommand {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractProjectorCommand.class);
@@ -330,8 +330,24 @@ public abstract class AbstractProjectorCommand<TKey, TValue, TOutput> extends Sm
      * @param <K> Key type
      * @param <V> Value type
      * @return a dead letter sink, which may be null to indicate none configured.
+     * @deprecated Use {@link #prepareDeadLetterSink(Object)} instead as not specifying the proper blank value for a DLQ
+     * may cause the DLQ to block an application
      */
+    @Deprecated(since = "1.6.0", forRemoval = false)
     protected <K, V> Sink<Event<K, V>> prepareDeadLetterSink() {
+        return null;
+    }
+
+    /**
+     * Prepares the dead letter sink, if any, where erroneous output from the projector is written.
+     *
+     * @param dlqBlankValue Custom DLQ blank value to use if the projection needs to replace the original value when
+     *                      sending to the DLQ because the event is too large to send as-is
+     * @param <K>           Key type
+     * @param <V>           Value type
+     * @return a dead letter sink, which may be null to indicate none configured.
+     */
+    protected <K, V> Sink<Event<K, V>> prepareDeadLetterSink(V dlqBlankValue) {
         return null;
     }
 
@@ -350,7 +366,7 @@ public abstract class AbstractProjectorCommand<TKey, TValue, TOutput> extends Sm
             try {
                 future.get();
             } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
+                Thread.currentThread().interrupt();
             } catch (Exception e) {
                 // Ignored, just trying to ensure that the driver has finished before we allow the JVM to exit
             }